feat: use join table inheritance
File, Video and Text entities inherit properties from Resource entity.
Resource entity will create a "resources" table with columns for each of
its properties. Each subclass will create a table with columns for each
of its own properties and an "id" column which is a foreign key related
to the "resources" table.